requirements.txt
# Discord
discord.py[voice]
dpy-ui

# Color Stuff
colorama
termcolor

# Database
pymongo
dnspython

# Misc
favicon

# VSCODE
pylint
autopep8